fix: H2-specific SQL-г H2 + PostgreSQL дэмжих ANSI SQL-р солих
AUTO_INCREMENT болон DOUBLE нь H2-д ажилладаг ч PostgreSQL-д алдаа өгнө.

Өөрчлөлтүүд:
- INT PRIMARY KEY AUTO_INCREMENT
    → INT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY
  (ANSI SQL стандарт — H2 2.x ба PostgreSQL 10+ дэмжинэ)
- DOUBLE → DOUBLE PRECISION
  (PostgreSQL-д зөвхөн DOUBLE PRECISION хүлээн зөвшөөрнө)
- cafes хүснэгтэд lat/lng баганыг CREATE TABLE-д шууд нэмсэн
  (ALTER TABLE migration мөрүүдийг устгасан)

+++ src/main/java/mn/edu/num/cafe/infrastructure/persistence/JdbcMenuRepository.java
@@ -35,7 +35,7 @@ public class JdbcMenuRepository implements IMenuRepository {
                     id        INT          PRIMARY KEY,
                     name      VARCHAR(255) NOT NULL,
                     category  VARCHAR(50)  NOT NULL,
-                    price     DOUBLE       NOT NULL DEFAULT 0.0,
+                    price     DOUBLE PRECISION NOT NULL DEFAULT 0.0,
                     available BOOLEAN      NOT NULL DEFAULT TRUE
                 )
                 """;
